third_party/WebKit/LayoutTests/http/tests/devtools/editor/text-editor-ctrl-d-1-expected.txt

This test verifies Ctrl-D functionality, which selects next occurrence of word.

function wordData() {
    return {
        original: $(".entry.original > .input").text(),
        translation: $(".entry.translation > .input").text(),
        tags: $(".active-tags > .tagcloud > .tag").toArray().map(function(value) { return value.textContent; })
    };
}

function submitWord(url) {
    var stub = new App.Stub($(".content"));
    $.post(url, wordData())
    .done(function() {
        var callback = $("meta[data-callback]").attr("data-callback");
        if (callback) {
            window.location = callback;
        } else {
            stub.success();
            $(".entry.original > .input").text("").focus();
            $(".entry.translation > .input").text("");
        }
    })
    .fail(function(obj, err, errDescr) {
        stub.failure("Error: " + errDescr);
    })
}


Running: testNextFullWord
'function': 3

Running: testCaseSensitive
'word': 2

Running: testOccurrencesOnTheSameLine
'original': 3

Running: testUndoLastAddedSelection
'original': 2

Running: testUndoSelectionPreservesFullWordState
'text': 3
Last selection: {"startLine":17,"startColumn":42,"endLine":17,"endColumn":46}

Running: testUndoSelectionPreservesPartialSelection
'text': 3
Last selection: {"startLine":4,"startColumn":96,"endLine":4,"endColumn":100}

Running: testTwoCloseWords
'text': 4
